requirements.txt
torch
numpy
matplotlib
pandas
python-dotenv